Carousel "Link to None" bug fixed
The primary issue here is that if you had Carousel enabled, and then created a gallery (or already had a gallery) and set it to Link To "None", then clicking any of the images in the gallery on the front end would fire the Carousel but it would be blank and not function properly or even be closeable.
The reason for this is that the Carousel JavaScript relied on a significant number of data attributes that were added to the image via the wp_get_attachment_link() function. Therefore these attributes were of course not added if the gallery was set to Link To "None".
I believe the reason for the oversight was that core didn't used to allow you to link a gallery to none. It was added about 5-6 months ago I think (about the time this problem was first reported.)
The fix for this has therefore turned out to be quite simple. Instead of adding the data attributes via wp_get_attachment_link(), they are now added via wp_get_attachment_image_attributes() which makes a lot more sense anyway. This means that even if you set your gallery to link to none, the carousel will still function properly (if it's enabled).
I hope this makes sense!
Props @jacklenox

This commit also includes a version bump for a js file -- the changes were exclusively upstream and not needed in Jetpack.