Multisite: Correct module slugs used on Network Settings screen.
Previously the module slugs were derived by running the module name through a few string manipuations, but this resulted in incorrect slugs in some cases. For instance, it resulted in "shortcodeembeds" instead of "shortcodes", and "contactform" instead of "contact-form".

Because of this, the modules failed to be activated when new sites were created because the incorrect slug was passed to activate_default_modules(), and it couldn't verify that the associated module file existed.

This commit gets the canonical slugs from get_available_modules() instead. As a result, list_modules() and its associated unit test are no longer needed.