Open Graph: og:url set improperly with previously static homepages
Props @blobaugh

og:url is set off of get_option( 'page_for_posts' )
If get_option( 'page_for_posts' ) returns 0 the home url is used, otherwise the id of the page set is used. This is the only check for a static homepage, however it does not work if the site admin changes back to posts for the home page. page_for_posts will still be set, thus causing og:url to show the wrong url.
The fix for this is to first check to see if the static home page option is set with get_option( 'show_on_front' )