Remove default post types for showing share links
Default post_types for showing share links were being set on line 204 but then were being overwritten. This updates the default to an empty array (which is the default when the global options have not been set) but now, the show parameter can be updated by the 'sharing_default_global' filter without being reset to an empty array.