MySQL: Make sure to define an ordering when using LIMIT
Using LIMIT without a defined ordering (as specified by an `ORDER BY`
clause) is not safe for a statement that modifies the table because
the affected result set is arbitrary and undefined, and so a different
result set may be affected on each slave server.

See https://dev.mysql.com/doc/refman/5.1/en/replication-features-limit.html

Reported by Abdulaziz AlJassar at http://wordpress.org/support/topic/mysql-warning-statement-may-not-be-safe-to-log-in-statement-format

Fixes #160.