Facebook Like Box: Fix default and minimum width.
In 4a3d04a5dafe401a2, the default and minimum width of the widget were
changed to reflect the Facebook documentation. However, this caused some
problems.

I have tested and it turns out the documentation is not true. Despite
the documentation claiming the widget has a minimum width of 292, in fact
a lower width will be respected. The actual minimum width appears to be
around 150 or so, but I will just set it back to 0 as it was before.

Reported by Elizabeth Urello (eurello).

Merges r90761-wpcom.