Custom CSS: revert r91252, as problem does not lie with output escaping, so change only masks the problem.
Fix addressed issue in output, but is not where the issue originates. Problem actual lies with the restoration action, as the handling found in `Jetpack_Custom_CSS::save()` doesn't fire for the restoration. To properly fix, elements should be abstracted and applied to the CSS during restoration.

Reopens #5894.

Merges r91617-wpcom.