newdash usercentric posts: This commits our work-in-progress code for the usercentric posts page
This commit should not impact anything in production, other than to create a new /posts page that is enabled for tritonites, which will have different styles applied, taken from the newdash-static repo. If this commit affects anything else, it should be reverted.

I am adding a temporary rest api endpoint for /me/posts which uses several mysql queries to pull posts from all blogs user is a member of. This endpoint will be rewritten to make use of elasticsearch when some backfilled data becomes ready. Also adding a /me/comments endpoint, but it's just a stub for now... little more than a copy of the /me/posts endpoint. Adding in styles from the newdash-static github repo, and reorganizing the /h4/tabs/functions.php file to separate the style and script enqueue statements so we can more easily isolate the css for the usercentric pages.

Merges r92695-wpcom.