Add a `wpml-config.xml` file.
This lets WPML recognize which of our options are strings that may need
to be translated.

A share feature option is a WP option, so to make it translatable we
need to add it to wpml-config.xml file. When plugins use get_option,
they read values from the wp_options table, WPML can filter these calls
and provide translation to the values of these options.

A greeting text option is a WP option, so to make it translatable we
need to add it to wpml-config.xml file. When plugins use get_option,
they read values from the wp_options table, WPML can filter these calls
and provide translation to the values of these options.

Props: WPML.org