Infinite: Only load new pages when within two screens of the bottom.
The current infinite scroll code loads a new page when the browser is
within 25% of the bottom of the page. On very long IS pages, this leads
to loading an increasing number of unnecessary pages, as the bottom 25%
of the page increases in size as the user loads more and more pages.

This patch makes IS less aggressive in loading new pages. Two screens was
chosen as a starting point and is obviously open to tweaking.