Multisite: Hides main blog from the Jetpack list
The main blog *must* be connected in order for Jetpack Multisite to be useful.
The main blog was included in the list of blogs displayed at Network Admin > Jetpack > Sites.
This let to some confusion amongst users when bulk deactivating sites and not unchecking the
main blog site. It would look like everything had been disconnected. This patch removes the
main site from the list (It *has* to be connected to see the list anyways!).

Fixes and closes 271