Update our Sass build process
It will now use our internal PHP Sass Compiler (that happens to be
bundled with the Custom CSS module), so there's one less external
dependency to worry about needing to install and have configured
identically.

Also renames the file, as it’s not using CSSTidy any longer, and uses a
custom style generator class for the PHP Sass to match core CSS Styles.