Add merge-data.json
This file will keep track of the last time each common file
was the same in both wpcom and Jetpack. This file will be
generated and updated by the new build script on wpcom, but
it is be safe to modify the entries in it by hand as well
(other than the comment at the top).