Carousel: Make slide titles visible in all viewport sizes
Before this change the slide titles would be cut off or missing on iPhone and
iPad in landscape orientation.

The position of the slide title was calculated as a percentage of the current
window height. When the window is short, the slide title doesn't fit.

This change solves the problem by measuring the height of the title and making
sure the position will be able to fit the entire thing. For large screens, we
stick with the existing proportion to preserve the original styles.

This change also includes some styles that are specific to iPad in landscape
mode to make things fit better.