Carousel: Switch slide when pressing back/forward buttons
Before this change, the browser's back and forward
buttons would update the URL with the ID of the
desired slide, but the UI would not actually
update.

The problem was that the `open` method was getting
called for `onhashchange` events, but it would do
nothing if the carousel was already activated.
This patch adds a new method that will switch the
slide even if the carousel is already open.