Centralize whitelisting for domains used by Jetpack services
Commit e8d9e676b4ee7d558eed6537dcdd63e8787bf18a introduced an additional two places
where a call to allowed_redirect_hosts is used and one additional function duplicating
the domain whitelisting efforts. This commit removes the need to repeatedly call and load
the whitelisted domains and centralizes it into one easy to manage method in the Jetpack
class thus removing the possibility for even more potential code duplication.