WPCOM: Shortcode: update from bandcamp on their shortcode.
Notes on the change from Neal Tucker ( ntucker@august20th.com ) @ Bandcamp:

We've added the ability to specify both the 'album' and 'track' parameters at once, to allow specifying a specific track to queue up on an album player.  This required a small amount of shuffling since there was an assumption that it would be either/or.  I also noticed that somehow, HTTPS handling got lost at some point, so I opted to remove the scheme from the URL entirely, leaving it as "//bandcamp.com/..."  Last time we had a conversation about that, the rationale for not using that approach was a windows XP issue, and it seems safe to ignore that now.  If you disagree, let me know and/or feel free to just change it.

Merges r95952-wpcom.