Tiled Gallery: Image dimensions to style attribute
Added image height and width attributes to a style attribute so that jQuery width and height accessors can access them before the image is loaded. Image height and width properties are mangled before rendering, so setting them doesn't make the browser reserve the required space before the image is loaded.