Fixed left nav & smoother mobile scrolling
- Made main wp-admin left nav fixed again
- Also resolved jerky scrolling in iOS

The one downside is that the little marker arrow in the “Jetpack” menu
item in the left nav is white which doesn’t look as good against the
green header area on the right.  It’s not too distracting though, and
is better than changing up the left nav so that it’s absolutely
positioned just on the JP admin pages.