REST API Edit Posts
We shouldn't default to 'post' if no type is passed. In editing, if something is omitted it means we want it to stay the same.
(bug introduced in [96553]). This was breaking things like the DMCA tool.

Also adds a test to make sure the type stays the same (so the bug can't regress).

Merges r96732-wpcom.