REST API: add new query paramater for filtering the fields returned
This reverts [96768], and thus un-reverts [96761].

This new query paramater allows you to specify which fields to return with endpoint responses by passing an additional comma-separated list of fields as part of the query.

For endpoints that contain one of the following objects: categories, comments, connections, domains, groups, likes, media, notes, posts, services, sites, suggestions, tags, themes, topics, users -- it will filter the contents of that object rather than the "parent" object. Otherwise, it filters the main returned object.

This was originally committed in [96768] however it broke notificaitons in the iOS app due to the notifications endpoint having an existing fields query paramater. The problem was not discovered due to the lack of tests for the notifications enpoint.

This improved commit makes sure that both methods for filtering the API responses are compatible with each other, and that the notifications endpoint functions properly. This also refines the tests used for the notifications endpoint as well as those for the filtering specifically.

This also removes an extra debug that was left in one of the tests.

see https://[private link] for project details

Merges r96808-wpcom.