REST API: Fix some post type/author assignment bugs
In [96732] a post type assignment bug was fixed. However, this now meant that cap checks for editing the author of a post always failed. This commit cleans up the post type and author assignments and the checks involved. It also fixes the `can_edit_post_and_assign_new_author` test which was failing while also passing all the other tests.

Merges r96824-wpcom.