REST API: remove temporary whitelist of VIP CPTs that are allowed in the API and always allow posts which are freshly pressed.
In [96828] I introduced a function for checking if a post is freshly pressed. This is now used in the REST API to bypass post type checks for the freshly pressed endpoints as they fail to return freshly pressed posts that are CPTs (e.g. on VIP sites) otherwise. Previously we had to whitelist every post type manually which was tedious and prone to error. This solves the problem in a more permanent fashion.

Merges r96833-wpcom.