Fixed error with setcookie() on bulk deactivate
On bulk module deactivate the value passed to setcookie() was an array. It needs to be a string.
Added in a check to ensure it is a string. @georgestephanis take a look and let me know if we
need to run through the array in a loop or if this check is ok.