First Pass at Manage architecture.
First pass: Notifying users of outdated plugins/themes, and linking
them to their dashboard to update them.

Or at least endpoints that pass that data back for display via the
WordPress.com Dashboard or the JSON API.

While this module is tentatively slated for auto-activation,
activate/deactivate endpoints will probably be made opt-in when
uncommented.