Contact Form: Switch the method by which we determine what numeric index new options should be given in the wizard.
The other method didn't account for a difference in how the fields are built when the field is brand new vs. when you're editing a form that already existed. In the former case, there's an element with a different that wasn't being included in our count when the field was brand new and which accordingly also necessitated the check for zero length that I here delete. Now we latch onto a different classname that selects both the first option and any subsequently added ones.

See http://[private link]

Merges r98059-wpcom.