Contact Forms: Prefix all fields with a number.
Prefixing all fields by a running number let's us keep the current simple
storing format while supporting having multiple fields of the same name and
type. These prefixes are removed in the admin view, as well as the success
message (though only if the prefix exists to not break previous feedback in
admin).