Contact Forms: Add `$all_fields` to subject filter
The `contact_form_subject` filter now receives an additional argument
`$all_fields` which is an array with the fields' label names as keys, and the
fields' values as values.

Refactored the hook for token replacing in the subject to use the passed field
value array instead of fetching the current form.