Jetpack Admin: Setting $instance private
Setting to private to block external access to the $instance property. By
having it set to public anything can come by and swap the object and Jetpack
could be sending data somewhere else or simply broken. This protects access.
If access to the object is required it can be reached with Jetpack_Admin::init()