Youtube Shortcode: Don't bother trying to show the shortcode in the admin.
Since we apply wp_kses_post with a later priority than this on comment_text anyway, the iframe gets stripped, and comments that have a url on its own line do not display the url in the comment admin area.

I looked a little bit into finding a clean way to apply the shortcode so that the videos would embed in the admin. Because we have to run the shortcode early to avoid make_clickable and because wp_kses_post fires with a priority of 10 in default-filters.php, I didn't really see a clean way. I guess we could remove wp_kses_post and re-add with a priority of 0 or something, but I didn't like the idea. If it needed to run super early, core would run it super early, I figure.

See http://[private link]

Merges r99419-wpcom.