SSO: Logout does not show login form when using auto-forward
When a user logged out they were being presented with the local login form. Users
should never be able to access this form. Now on logout the login form will be
hidden and the user not forwarded to WordPress.com until they revisit the login page.
The form is hidden instead of the forward happening because the log out only affects
the local site and the user is still logged into WordPress.com. As such the user would
be stuck in an infinite loop attempting to logout if the forward was performed. Simply
hiding the local form and showing the login with WordPress.com button should do the
trick quite nicely.