Fix landing page featured module modals.
The landing page featured module modals didn't show any content when using
translations, because the contents were found using the module name string
embedded untranslated in the HTML to find the matching module from a list of
modules.

By using the module string (like custom-css) instead of the name of the module
we can find a match regardles of translation or escaping done by json_encode
for example on some Spanish translations.