Custom CSS: don't bother checking for a migrated CSS revision of the blog never had Custom CSS before it was migrated. Fixes #6046.
The issue reported was Sass wasn't compiling correctly in the front-end (saved revision). Turns out it was never stored properly because only "migrated" revisions store a "filtered" CSS string, which is where pre-processed CSS is stored.

See also #5957 -- although that will be a separate issue to solve.

Merges r101016-wpcom.