Likes: update the Settings > Discussion > "Someone likes one of my posts" toggle to appear enabled by default.
This feature is already enabled by default, but this option made it appear as though the feature is disabled
in certain situations. On blogs where the settings on the Discussion page haven't been updated (e.g. a new
blog), the absence of a saved value for this option was causing the Likes notifications to be sent -- even
though this option appeared disabled on the admin page.

https://[private link]

Merges r101245-wpcom.