Publicize: allow html entities to be part of the messages sent to services without allowing actual html tags
Standardize the way we sanitize the messages sent to services into a single method `Publicize_Util::sanitize_message`, which first removes `script` and `style` tags, then runs the content through kses, trims whitespaces and decodes any left over entities.

This allows users to have hearts and other symbols in their publicize messages. This is a redux of [101294] which was reverted in [101468] because it was causing html to be displayed in publicized posts.

The following code has been tested to ensure that the problems from before are no longer present.

see https://[private link]
see https://[private link]

Merges r101489-wpcom.