Add a title/description extractor and compiler.
For v3.2 and on, this will let us translate module headings without a
custom parser, and will work more gracefully with the WordPress.org
plugin translation glotpress install.

Please don’t commit the results to the repo yet, as we don’t want new
strings in our current wpcom glotpress for 3.1 this close to release.