Refining the setting of site_tag when twitter:creator is present.
This is an attempt to close #935. I added a class method that
will take in the current value for site_tag and og_tags and
return a new value for site_tag if twitter:creator is set in
og_tags. If nothing is set, it returns the original value of
site_tag.