Shortcodes: Fix [googlemaps] shortcode reversal false positive.
`jetpack_googlemaps_embed_to_short_code()` is intended to only reverse URLs in
the format of `https://maps.google.com` or `https://google.com/maps/`, but it
is matching `google.com` any domain.

For example, the following tag will be reversed into a `[googlemaps]` shortcode
even though it should not be:

<iframe src="https://docs.google.com/spreadsheets/d/19FacgevvN-GGS8SOlbMRVKkWDYSMDQEBfRmkwqaA9z8/pubhtml?widget=true&amp;headers=false"></iframe>

The problem is that the code mistakeningly tests for a `false` return from
`preg_match()` instead of `0` or `false`. Unlike `strpos()`, `preg_match()`
returns `0` if the pattern is not matched, and `false` if there is an error.
Because of this, the condition evaluates to true, and the function
doesn't returns early.

This bug was introduced in 7d5c5a77b25474798554fba4e8f72c853f96b507.