JSON API: Make sure datetimes are always well formed.
WordPress stores `date` and `date_gmt` fields for posts (same for `modified` and `modified_gmt`).

The API's date formatting code uses both of these values to generate a datetimestring including the local time and a timezone offset for that post.

For drafts, though, date_gmt is set to "0000-00-00 00:00:00", which is messing up the API's date formatting code.

Dates were being output like: "2014-02-14T09:21:51+17656209:21".

When the gmt field is missing like that, use the blog's timezone/offset option to calculate the correct offsite time.

http://[private link]

Merges r102805-wpcom.