Jetpack: Cache file_data in an option.
And index it by version.

This saves about 105 ms (it had been less previously, but when we
removed the prior `file_exists` checks, the `fopen` calls in
`get_file_data` went up.)

This does mean that we’ll probably need some sort of way to turn this
off when in active development, so changes we make to the module
headers actually get seen.  Maybe compare it on the admin side to the
real headers if `?page=jetpack` — unsure.