Resolve error from WP core being thrown in Jetpack tests
When setting the siteurl database option WordPress core attempts to rewrite the .htaccess file
When this happens WordPress attempts to determine the absolute file system path, however when
doing so WordPress has a little bug that causes it to be confused and claim the WP install
lives at / instead of the actual file system path. Added a stupid hack to prevent WP core from
breaking the legitimate test of Jetpack.

Related #1042
Related #1021