REST API
New endpoint for loading a theme's post formats. The current method (now deprecated - and will be removed in a vesioning) doesn't work properly.
This new endpoint uses [104371] to load a theme context so we always get the right set of formats.
This change is going to be used to fix https://github.com/Automattic/beep-beep-boop/issues/48 in the interim editor.

Merges r104372-wpcom.