Add a set of classes for unit testing the shortcodes.
There are a few small issues in the shortcodes, both related to
functionality (the is_numeric test to allow ids in flickr) and
to allow unit testing (not assuming a global $post is defined).

Also, there are a lot of lines added here, but the shortcode
unit test classes all share the same first two functions to make
sure that a) the shortcode is properly defined, and b) that the
shortcode is caught and dealt with by some callback.

Some classes have more tests than others for now. If this is fine,
however, the files should be a good base for other contibutors
(or me!) to add a more diverse range of tests.