ShareDaddy: Make sure we're sending people to the https link.
As this is an undocumented sharing link, there may be back-end api changes
without notice.  This seems to have been one of them.

When sending folks to the http:// version of this url, they were experiencing
a redirect loop -- it would cycle between

http://www.linkedin.com/cws/share?isFramed=false&url=http%3A%2F%2Fexample%2Ecom%2F%3Fp%3D1&token=

and

https://www.linkedin.com/uas/connect/user-signin?session_redirect=http%3A%2F%2Fwww%2Elinkedin%2Ecom%2Fcws%2Fshare%3FisFramed%3Dfalse%26url%3Dhttp%253A%252F%252Fexample%252Ecom%252F%253Fp%253D1%26token%3D

Basically, the login would set the cookie for https only, but the sharing link wouldn't see the cookie as it's httpsonly, and so redirect back to the login page, which said "wtf, you're already logged in, just go to your redirect!" -- and the loop continues.

Anyway, by changing so we only send to the https version, this should be resolved.

Merges r104610-wpcom.