Site Icon: JSON API Update
Site icon as different functions from .com so we needed to update the
returned api end points.
We are using photon url because this way we can get https urls.
Right now we are not generation any ico files. Photon doesn’t support
them so we are returning a png file.

The dechex conversion is done to bust the photon cache.