infinity.js : Resolve Divergence from Jetpack
This patch teaches the infinite scroll system how to pull DOM nodes out of
nodes outside the current viewport and stash them in a cache. It does this
by hooking into determineURL and walking all of the pages that were determined
to be hidden. It pulls all of the child nodes out of the page and stashes them
in a DocumentFragment. It also updates the min-height of the page wrapper to
match the old height when all the children were present.

When a user scrolls a page back into view, the children are reinserted into
the wrapper and the min-height and class marker are removed.

See https://github.com/Automattic/jetpack/pull/300

But also fixes : https://[private link]

The cache that was stored did not work with lazy load or any other plugin that fires on the 'post-load' event.
Since the post load wasn't fired every time the content was added or re-added to the dom.
Now we store the html also right before we take it away because it might have been altered by the post-load event.

Another  optimization that was done to fire the change url  in the same timer as the the refresh() since this will only create one times instead of multiple once.

Merges r106807-wpcom.