REST API: add embed/shortcode reversal endpoint
This endpoint allows the reversal of a third party embed code into a shortcode or single-line embed based on embed/shortcode reversals which we already have in WordPress.com/Jetpack.

It accepts the third party embed code as a receiving paramater and, if valid and properly reversed, returns the reversed shortcode/embed as well as the rendered result of that shortcode/embed and any accompanying scripts or styles needed for the render. This is similar to the existing embed/shortcode rendering endpoints.

As part of this a few small changes were made to the shared class for rendering embeds/shortcodes in the API. A few tests were also added for the reversals.

see https://[private link]

Merges r106903-wpcom.