REST API Media Endpoints (1.1)
Fix a few issues/bugs with the code: $insert was not defined as an empty array so attrs could get passed thru to other items. Also updated media_urls to associate errors with the correct index.
Finally made sure IDs are returned as ints like they are documented as.

Merges r107147-wpcom.