REST API Media Endpoints
v1.1 of the listing endpoint was using 'numberposts' to set the number of requested items (because an earlier version of the code was using get_posts which is accepted). However, numberposts does not work for WP_Query. posts_per_page needs to be used instead.
Two tests have been added to MediaTest to make sure this works correctly.

Merges r107344-wpcom.