Post flair/sharing: add the genericon for each service as a property of each respective class
I'm adding a REST API endpoint for `meta/sharing-buttons` which provideds the needed meta for managing the sharing buttons for sites. As part of this, I want to return the correct genericon to use for each one to make it easier to output previews using the API response.

Related: https://github.com/Automattic/calypso-pre-oss/issues/1329

Merges r107563-wpcom.

Conflicts:
	modules/sharedaddy/sharing-sources.php