Plugin Endpoint: Rethink autoupdates
Yesterday, I attempted to fix it so that when a calypso user chose to autoupdate a plugin,
that we would schedule a cron event to trigger the next autoupdate immediately. Today,
I thought of a better way. Now we will pass an action='update' parameter with the request
anytime we submit a request with autoupdate='true' - this will envoke the ::update() method
in our class, and we won't need to use cron at all.