REST API: Posts v1.1: Move comment- and ping-related properties to nested "discussion" object property
This is a backwards-incompatible change. `comments_open` and `comment_status` are not synonymous, and can be different if the user chooses to have discussion remain open for a specified number of days (at which point, `comment_status` is `open` but `comments_open` is `false`).

See: https://github.com/Automattic/beep-beep-boop/issues/156

Merges r109553-wpcom.