REST: add page handles to /me/posts and fix up encoding inside /sites/site/posts/ handles
Add page handles for paging to the /me/posts endpoint. They work in the same manner
as the newly added handles or /sites/$site/posts.

Also fix a bug in how we encode the value into the handle for /sites/$site/posts/.
Previously, we were url encoding the entire handle to stick it into the next page URL,
but we were not encoding the individual pieces. This meant that when we parsed the handle
with wp_parse_args, we were parsing a string that was build without the proper escaping.
Any value that contained an special character, like %, &, +, would end up parsing incorrectly.

The code now properly encodes the values, and then encodes the entire handle.

Merges r109758-wpcom.