Plugin Modify Endpoint: clear cache before updating
We were clearing the plugin cache within the foreach loop, which was optimal
for updating multiple plugins at once. But we've decided on calypso that it
is better to update one plugin per api call - it gives us better error reporting
and progress reporting. Thus we should clear the plugin cache before
starting the update process.