REST: switch paging handles over to just the value, instead of the full URL.
The full URL was too difficult to built reliably, as we include both parameters
that affect the query and parameters that affect how we output the response, like
callback and http_envelope. Additionally, every client I've written for the page
handles really has to extract the value from the URL and use it as a value anyway,
as clients for our API want a base URL and params, not a fully-formed URL.

See https://[private link]

Merges r109976-wpcom.