Bug fix: Don't indicate what modules activated
If the user doesn’t have a connected jetpack site or is in developer
mode. don’t show if the module is active. Since it shouldn’t do
anything but can cause confusion in the UI.
Since some boxes will be highlighted as activeed and others will not.
The user is not able to click the deactivate links.