Moved setScaledImageSrc function to proper place
setScaledImageSrc was hidden in an "If" statement. This is not a proper place for a function that leads to the main functionality. The position of important function inside if clause can mislead any future code analysis. The change is quite simple - get setScaledImageSrc return to variable and than check if this variable is true or false.