Moving all functionality of image analysis into setScaledImageSrc
When we are looping through all the images, we need to decide if the images are suitable for scaling. If they arent loaded yet, or they are marked as lazy load, or maybe failed to load scaled src in the past. All those conditions are being stated in isImageIsScalable. isImageIsScalable was called in zoomImages function - in the looping of the images. But there is no need to do that. We can move isImageIsScalable function to setScaledImageSrc, Now setScaledImageSrc is the only function that is being called in the image looping. if isImageIsScalable returns false, setScaledImageSrc will not continue.