REST API: Do not return malformed JSON for taxonomies when user is a contributor
When a user had the capability to edit a post but not to manage taxonomies, we
were returning malformed JSON for the GET /posts and GET /posts/$post_id endpoints.
This was due to passing the context along to format_taxonomy.

Instead, force the context to 'display' for categories and tags, since anyone
who can edit a post can view/add the existing taxonomies.

https://github.com/Automattic/io/issues/112

props @koke

Merges r110651-wpcom.