Jetpack Media Summary: prevent a post from returning a type of `gallery` when it doesn't have any images retrieved in the post.
I discovered that it was possible for a post's media summary to find a gallery but fail to extract the images from it, in which case the type would be gallery but no images would be found. This caused problems for Publicize which, for some services (like Path) expects there to be an image when the type is `gallery`. As a result we tried to send an empty to Path which caused Publicize to fail for that service for the affected posts.

This fix makes it so that if no images are extracted, the type is returned as `standard`. I'll see if it's possible to fix the image extraction bug in a separate commit.

Merges r111005-wpcom.