REST API: streamline multi_user check in /me/sites
This is in response to feedback about improving the /me/sites endpoint that came in through this post: https://[private link]. I've added an is_current_site_multi_user() method to the base API as suggested and am using a cache to speed things up a bit.

Merges r111107-wpcom.