Protect Module: main blog id
On activation, if we can't find the main blog id, that means we are
on multisite, and the main blog hasn't connected to Jetpack. If this
is the case, we will bail on getting the protect api key.

since there will be cases where getting the protect api key will fail,
we'll need to make a configuration page for this module that will give
more information, and give them a chance to retry. the next commits will
focus on the config page.