saving different option empty_trash_days instead of trash_folder_disabled
Couple things this does, now we're storing an int instead of a boolean which makes the check for the value against the shadow blog a little easier. Also, it's a little less magical, since this is the actual config setting that determines the value. I'm also switching to update_option since the pre_option approach wasn't syncing. And last I'm adding blogdescription to the synced options, since it won't otherwise sync if you have stats disabled.