Added Mock Jetpack Option to Sync options that don't have to live in the Database
If you want to sync options that don’t live in the database. Before you
had to:
Add the option to be synced to .com by calling the sync_options method.
Add a filter to pre_option_ to fake retrieve it from the db.
Trigger the change unless you want the synced by when
‘sync_all_registered_options’ method is called. Which is daily by
heartbeat and on plugin upgrade.

Now you can use the jetpack_mock_option function that lets you define
the jetpack option and the callback that would  return the output of
it.